+++ configs/configs.go
@@ -87,6 +87,7 @@ func initConfig() {
 	getOrSetDefaults("GHORG_BITBUCKET_APP_PASSWORD")
 	getOrSetDefaults("GHORG_SCM_BASE_URL")
 	getOrSetDefaults("GHORG_PRESERVE_DIRECTORY_STRUCTURE")
+	getOrSetDefaults("GHORG_OUTPUT_DIR")
 }
 
 // Load triggers the configs to load first, not sure if this is actually needed